# feat: Priority 2 UX Optimization - Web Worker Watermarking, Performance Monitoring & React.memo