--- a/.claude/ux-blockers.md
+++ b/.claude/ux-blockers.md
@@ -1095,3 +1095,4 @@ if (typeof OffscreenCanvas === 'undefined') {
 ---
 
 **Ready to proceed with implementation?** I can help execute any phase when you say "Proceed" with the phase number.
+r