--- a/package.json
+++ b/package.json
@@ -45,7 +45,8 @@
     "react-easy-crop": "^5.4.2",
     "react-router-dom": "^6.26.2",
     "tailwind-merge": "^2.5.2",
-    "tailwindcss-animate": "^1.0.7"
+    "tailwindcss-animate": "^1.0.7",
+    "web-vitals": "^5.1.0"
   },
   "devDependencies": {
     "@eslint/js": "^9.9.0",