--- a/src/components/product/PhotoCropper.tsx
+++ b/src/components/product/PhotoCropper.tsx
@@ -1,5 +1,5 @@
 
-import { useEffect } from "react";
+import { useEffect, memo } from "react";
 import { Card } from "@/components/ui/card";
 import CropperHeader from "./cropper/components/CropperHeader";
 import OrientationSelector from "./cropper/components/OrientationSelector";
@@ -18,7 +18,7 @@ interface PhotoCropperProps {
   onChangePhoto?: () => void;
 }
 
-const PhotoCropper = ({
+const PhotoCropper = memo(({
   imageUrl,
   initialAspectRatio = 1,
   selectedOrientation = "square",
@@ -131,6 +131,8 @@ const PhotoCropper = ({
       </div>
     </div>
   );
-};
+});
+
+PhotoCropper.displayName = 'PhotoCropper';
 
 export default PhotoCropper;