… src/utils/watermarkUtils.ts
… src/workers/watermark.worker.ts
… supabase/migrations/20251002000000_performance_metrics.sql